[ROS2] Refactored ROS2GemUtilities
Refactored ROS2GemUtilities to avoid compilation issues with the
ROS2.Static target and moved the CreateComponent function to the
RobotImporterUtils
